Modify the documentation to add link from original readme update the filemetadata to express filesize as long

+++ src/main/java/com/woven/movie_uploader/filehandler/FileMetadata.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 
 // Metadata file that can be converted to JSON.
 public record FileMetadata(@JsonProperty("fileid") String fileId, @JsonProperty("name") String name,
-                           @JsonProperty("size") int filesize, @JsonProperty("created_at") String createdAt,
+                           @JsonProperty("size") long filesize, @JsonProperty("created_at") String createdAt,
                            @JsonIgnore String contentType) {
 
     public FileMetadata {
